+++ index.d.ts
@@ -467,7 +467,8 @@ declare module 'pc-nrfconnect-shared' {
     // AppTypes.ts
 
     export type SourceJson = import('./src/utils/AppTypes').SourceJson;
-    export type AppJson = import('./src/utils/AppTypes').AppJson;
+    export type AppInfo = import('./src/utils/AppTypes').AppInfo;
+    export type AppVersions = import('./src/utils/AppTypes').AppVersions;
     export type PackageJson = import('./src/utils/AppTypes').PackageJson;
 
     // useHotKey.js