+++ scripts/nordic-publish.ts
@@ -13,10 +13,10 @@ import FtpClient from 'ftp';
 import semver from 'semver';
 import calculateShasum from 'shasum';
 
-import { AppJson, PackageJson, SourceJson } from '../src/utils/AppTypes';
+import { AppInfo, PackageJson, SourceJson } from '../src/utils/AppTypes';
 import checkAppProperties from './check-app-properties';
 
-interface AppInfo {
+interface LegacyAppInfo {
     ['dist-tags']?: {
         latest?: string;
     };
@@ -36,7 +36,7 @@ interface App {
     version: string;
     shasum: string;
     sourceUrl: string;
-    appJsonName: string;
+    appInfoName: string;
     releaseNotesFilename: string;
     iconFilename: string;
     isOfficial: boolean;
@@ -149,7 +149,7 @@ const packOrReadPackage = (options: Options): App => {
         shasum,
         sourceUrl: options.sourceUrl,
         isOfficial: options.deployOfficial,
-        appJsonName: `${name}.json`,
+        appInfoName: `${name}.json`,
         releaseNotesFilename: `${name}-Changelog.md`,
         iconFilename: `${name}.svg`,
         packageJson,
@@ -207,7 +207,7 @@ const downloadFileContent = (filename: string) =>
         });
     });
 
-const downloadAppInfo = async (name: string): Promise<AppInfo> => {
+const downloadLegacyAppInfo = async (name: string): Promise<LegacyAppInfo> => {
     try {
         const content = await downloadFileContent(name);
         return JSON.parse(content);
@@ -221,7 +221,7 @@ const downloadAppInfo = async (name: string): Promise<AppInfo> => {
     }
 };
 
-const updateLocalAppInfo = (appInfo: AppInfo, app: App) => {
+const updateLegacyAppInfo = (appInfo: LegacyAppInfo, app: App) => {
     const latest = appInfo['dist-tags']?.latest;
     if (latest != null) {
         console.log(`Latest published version ${latest}`);
@@ -263,9 +263,9 @@ const uploadFile: UploadLocalFile & UploadBufferContent = (
         client.put(local, remote, err => (err ? reject(err) : resolve()));
     });
 
-const getUpdatedAppInfo = async (app: App) => {
-    const appInfo = await downloadAppInfo(app.name);
-    return updateLocalAppInfo(appInfo, app);
+const getUpdatedLegacyAppInfo = async (app: App) => {
+    const appInfo = await downloadLegacyAppInfo(app.name);
+    return updateLegacyAppInfo(appInfo, app);
 };
 
 const downloadSourceJson = async () => {
@@ -303,7 +303,7 @@ const getUpdatedSourceJson = async (app: App): Promise<SourceJson> => {
         name: sourceJson.name,
         apps: [
             ...new Set(sourceJson.apps).add(
-                `${app.sourceUrl}/${app.appJsonName}`
+                `${app.sourceUrl}/${app.appInfoName}`
             ),
         ].sort(),
     };
@@ -311,10 +311,8 @@ const getUpdatedSourceJson = async (app: App): Promise<SourceJson> => {
 
 const downloadExistingVersions = async (app: App) => {
     try {
-        const appJson = <AppJson>(
-            JSON.parse(await downloadFileContent(app.appJsonName))
-        );
-        return appJson.versions;
+        const appInfoContent = await downloadFileContent(app.appInfoName);
+        return (JSON.parse(appInfoContent) as AppInfo).versions;
     } catch (error) {
         console.log(
             `No previous app versions found due to: ${errorAsString(error)}`
@@ -330,7 +328,7 @@ const failBecauseOfMissingProperty = () => {
     );
 };
 
-const getUpdatedAppJson = async (app: App): Promise<AppJson> => {
+const getUpdatedAppInfo = async (app: App): Promise<AppInfo> => {
     const versions = await downloadExistingVersions(app);
 
     const { name, displayName, description, homepage, version } =
@@ -343,22 +341,19 @@ const getUpdatedAppJson = async (app: App): Promise<AppJson> => {
         homepage,
         iconUrl: `${app.sourceUrl}/${app.iconFilename}`,
         releaseNotesUrl: `${app.sourceUrl}/${app.releaseNotesFilename}`,
-        latest: version,
+        latestVersion: version,
         versions: {
             ...versions,
             [version]: {
-                tarball: `${app.sourceUrl}/${app.filename}`,
+                tarballUrl: `${app.sourceUrl}/${app.filename}`,
                 shasum: app.shasum,
             },
         },
     };
 };
 
-const uploadAppInfo = (app: App, updatedAppInfo: AppInfo) =>
-    uploadFile(
-        Buffer.from(JSON.stringify(updatedAppInfo, undefined, 2)),
-        app.name
-    );
+const uploadLegacyAppInfo = (app: App, appInfo: LegacyAppInfo) =>
+    uploadFile(Buffer.from(JSON.stringify(appInfo, undefined, 2)), app.name);
 
 const uploadSourceJson = (sourceJson: SourceJson) =>
     uploadFile(
@@ -366,10 +361,10 @@ const uploadSourceJson = (sourceJson: SourceJson) =>
         'source.json'
     );
 
-const uploadAppJson = (app: App, appJson: AppJson) =>
+const uploadAppInfo = (app: App, appInfo: AppInfo) =>
     uploadFile(
-        Buffer.from(JSON.stringify(appJson, undefined, 2)),
-        app.appJsonName
+        Buffer.from(JSON.stringify(appInfo, undefined, 2)),
+        app.appInfoName
     );
 
 const uploadPackage = (app: App) => uploadFile(app.filename, app.filename);
@@ -417,15 +412,15 @@ const main = async () => {
         await connect(config);
         await changeWorkingDirectory(options.sourceDir);
 
-        const appInfo = await getUpdatedAppInfo(app);
+        const legacyAppInfo = await getUpdatedLegacyAppInfo(app);
         const sourceJson = await getUpdatedSourceJson(app);
-        const appJson = await getUpdatedAppJson(app);
+        const appInfo = await getUpdatedAppInfo(app);
 
         await uploadChangelog(app);
         await uploadIcon(app);
         await uploadPackage(app);
+        await uploadLegacyAppInfo(app, legacyAppInfo);
         await uploadAppInfo(app, appInfo);
-        await uploadAppJson(app, appJson);
         await uploadSourceJson(sourceJson);
 
         console.log('Done');